refactor (move input validate method)
